+++ Assets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -44,6 +44,7 @@ MonoBehaviour:
   - ApplyItem
   - ApplyItemRPC
   - PlayTimeLineRPC
+  - SpawnItemRPC
   DisableAutoOpenWizard: 1
   ShowSettings: 1
   DevRegionSetOnce: 1
